fix: apply edits to line and print once
Prior to this change, if guid was found in an input line more than once
i twould print the line once for each change, and only include the
change from that guid capture result, rather than that plus the previous
ones.

This change uses a mutable string to store the changes from g -> u and
then prints the result at the end once.